V1.1.0 contrib bug fixes (#383)
HElib 1.1.0 Beta 0, August 2020
===============================
(tagged as v1.1.0-beta.0)

* Throw `LogicError` on logic failure
* Fix PAlgebra equality operator
* Initialize `cost` for UpperMemoEntry and LowerMemoEntry to 0
* Change `found`  to bool in function makeMask
* `unsigned long` cannot be guaranteed to be `size_t`
* Typo in comments

+++ src/Ptxt.cpp
@@ -861,7 +861,7 @@ long Ptxt<Scheme>::coordToIndex(const std::vector<long>& coords)
 {
   const PAlgebra& zMStar = context->zMStar;
   assertEq<LogicError>(coords.size(),
-                       static_cast<unsigned long>(zMStar.numOfGens()),
+                       static_cast<std::size_t>(zMStar.numOfGens()),
                        "Coord must have same size as hypercube structure");
   long index = 0;
   // Convert the coordinates into its corresponding index by computing the
